path/minor formatting changes
Updated the includes statement for all files in the examples folder so they have the correct path to the connction script. Updated where to find the DB file for create_db.php to be the examples folder; updated where connction.php should look for this db to the examples folder. when actually using connection.php, this will have to be changed. For now, you can run the examples straight from the examples folder to see how everything works.