added mail function + test code
mail function added; it is a one-line piece of code. added test code that displays what the user typed in. could be useful if we want the feedback to show the user what they just submitted. in order to set this up on server where canvas is installed, we need mail server info and have to update php.ini.