re-created mail function using phpmailer
instead of using mail function directly, used PHPMailer, which handles the more nuanced aspects of constructing and sending an email. The code is a working test; right now it lacks pulling in the user's email from database. i added a from field so that you can test and see how it works. to test, create a local email server using something like papercut (link to papercut is in comment in file). i tried using mark's email server but it only worked for me once, and was non responsive after that. once we get more info on how email is set up on the canvas server, this can be added to the file.