added load screen
created a load modal dialog box. it populates with the name of quizzes, much like the drop down did before. it creates a button for each survey, then passes the surveyID to load function. adjusted survey_ids.php so that it has error correction. it returns a surveyID value of 0 if table empty, and thorws an error if table not found