updated php files and added new db
added a new db called db.sqlite. this has extra tables to support users and user ids, and a linking table to join user, answer and quiz. also fixed some logic with saving quizzes so all keys in the json for a quiz is now quoted properly (or should be)