fixed editor modal not opening
re organized the script and link tags for js and css in header and footer so that the editor's modal works again. whenever clicking edit on a question, the modal woudl not show. the console showed an error. after checking online, i found that bootstrap's js has to load before its css to avoid this. also we had 2 jquery imports, i removed one.