pull_json_data.php can return all user data
i added some logic to the json pulling script so it only accepts positive integers and the keyword all. this info has to be passed via an ajax method using POST. add {'ID': and some value or variable (can be string numbers) and it will populate the results.json file with the results.