added functionality to tip_viewer
in preparing for a quicksave feature, i added logic to tip_viewer.js which looks up a users ID (it will be taken from the [OAUTH_ID] once I can test it) and will look for the currently assigned tip for that user. if there are no saved answers, it starts a fresh assessment. if there are saved answers and the answers are marked incomplete, they can pick up where they left off. if the answers are marked complete, it prompts them that they have already completed this assessment.