tip saving updates db tables properly
saving survey answers now updates not only the answers table (either inserting new row or updating existing row) it also updates the joining table. updating the joining table is critical for being able to update existing data in the answers table. this is what allows a survey that was quick saved to display answers to questions the user already answered.